# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gardina API
  version: 1.0.0
  description: >-
    Multi-tenant SaaS ERP for curtain/drapery salons. Single source of truth for
    the web and iOS clients. All resource routes are tenant-scoped via the JWT
    (the token carries `organizationId`); clients never pass tenant ids explicitly.


    AUTH: JWT Bearer. Access token ~7d, refresh token ~30d, bcrypt passwords.


    READ-ONLY MODE: when the organization's subscription is expired/past_due/canceled,
    write operations return 402 with `code: SUBSCRIPTION_READ_ONLY`. The `/auth` and
    `/billing` route groups are whitelisted.


    RATE LIMITS: global 100 req / 15 min per IP on `/api`. Stricter per-route limits:
    login 5/15min, password-reset 5/15min, register-salon & register 3/h, refresh 30/15min.


    RESPONSE ENVELOPE: most endpoints return `{ success: true, data: ... }`. Errors
    return `{ success: false, error: "..."}` or `{ error, message, timestamp }` from
    the global handler. Shapes vary by controller; this contract documents the common
    cases.
servers:
  - url: https://api.gardina.kz/api
    description: Production
  - url: http://localhost:5000/api
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /deals:
    get:
      tags: [Deals]
      summary: List deals
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/DealStatus' }
        - name: designerId
          in: query
          schema: { type: string, format: uuid }
        - name: clientId
          in: query
          schema: { type: string, format: uuid }
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Deal list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Deal' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Deals]
      summary: Create a deal
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DealCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DealResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /deals/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Deals]
      summary: Get a deal by id (with events/timeline)
      responses:
        '200':
          description: Deal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DealResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [Deals]
      summary: Delete a deal
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/DealStatus' }
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Deal' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Orders]
      summary: Create an order (deal)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DealCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DealResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '402': { $ref: '#/components/responses/ReadOnly' }
  /orders/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Orders]
      summary: Get an order by id
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DealResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Orders]
      summary: Delete an order
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /clients:
    get:
      tags: [Clients]
      summary: List clients
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Client list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Client' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Client' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '402': { $ref: '#/components/responses/ReadOnly' }
  /clients/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Clients]
      summary: Get a client by id
      responses:
        '200':
          description: Client
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Client' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Clients]
      summary: Update a client
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Client' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Clients]
      summary: Delete a client
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ────────────────────────────────────────── MEASUREMENTS ───────────────────
  /measurements:
    get:
      tags: [Measurements]
      summary: List measurements
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/MeasurementStatus' }
        - name: designerId
          in: query
          schema: { type: string, format: uuid }
        - name: clientId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Measurement list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Measurement' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Measurements]
      summary: Create a measurement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MeasurementInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Measurement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '402': { $ref: '#/components/responses/ReadOnly' }
  /measurements/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Measurements]
      summary: Get a measurement (with windows)
      responses:
        '200':
          description: Measurement
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Measurement' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Measurements]
      summary: Update a measurement (and its windows)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/MeasurementInput'
                - type: object
                  properties:
                    windows:
                      type: array
                      items: { $ref: '#/components/schemas/MeasurementWindowInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Measurement' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /proposals:
    get:
      tags: [Proposals]
      summary: List proposals
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ProposalStatus' }
        - name: clientId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Proposal list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Proposal' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Proposals]
      summary: Create a proposal
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProposalInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Proposal' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '402': { $ref: '#/components/responses/ReadOnly' }
  /proposals/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Proposals]
      summary: Get a proposal by id
      responses:
        '200':
          description: Proposal
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Proposal' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Proposals]
      summary: Update a proposal
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProposalInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Proposal' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Proposals]
      summary: Delete a proposal
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ──────────────────────────────────────────── USERS ────────────────────────
...
  /users:
    get:
      tags: [Users]
      summary: List all users (admin, manager)
      parameters:
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/UserRole' }
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{id}:
    get:
      tags: [Users]
      summary: Get a user by id (admin, manager)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /payments:
    get:
      tags: [Payments]
      summary: List payments with filters
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: dealId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/PaymentStatus' }
      responses:
        '200':
          description: Payment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Payment' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Payments]
      summary: Create a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dealId, amount]
              properties:
                dealId: { type: string, format: uuid }
                amount: { type: number, format: float }
                method: { type: string }
                type: { type: string, enum: [prepayment, final_payment] }
                note: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Payment' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '402': { $ref: '#/components/responses/ReadOnly' }
  /payments/{id}:
    delete:
      tags: [Payments]
      summary: Delete a payment
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '402': { $ref: '#/components/responses/ReadOnly' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ──────────────────────────────────────── NOTIFICATIONS ────────────────────
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List the current user's notifications
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: unread
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Notification' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Notifications]
      summary: Create a notification (admin/system)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, type, title]
              properties:
                userId: { type: string, format: uuid }
                type: { $ref: '#/components/schemas/NotificationType' }
                title: { type: string }
                message: { type: string }
                actionUrl: { type: string, format: uri }
                metadata: { type: object, additionalProperties: true }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Notification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

    # ── Enums ───────────────────────────────────────────────────────────────
...
    DealStatus:
      type: string
      enum:
        - lead
        - measurement_scheduled
        - measurement_done
        - proposal_sent
        - proposal_accepted
        - contract_signed
        - in_production
        - ready_for_installation
        - installation_scheduled
        - installed
        - completed
        - cancelled
...